safdfsda

+++ main.py
@@ -15,7 +15,7 @@ quantityinput=int(input("Please select a quantity level from 1-100\n"))
 tryinput=int(input("Please enter how many tries you'd like. \n"))
 
 quality=float(1+float(qualityinput)/100)
-quantity=int(quantityinput)
+quantity=float(quantityinput)
 
 additionalrollscount=0
 additionalrollschance=10000.00
